Remove mapiauth state management domain, and move everything to the main state domain
