+++ src/components/Auth/MAPI/MapiLoginPage.tsx
@@ -1,13 +1,14 @@
 import MapiLoginButton from 'Auth/MAPI/MapiLoginButton';
 import MapiLoginStatusText from 'Auth/MAPI/MapiLoginStatusText';
+import { push } from 'connected-react-router';
 import { FlashMessage, messageTTL, messageType } from 'lib/flashMessage';
 import MapiAuth from 'lib/MapiAuth/MapiAuth';
 import React, { useCallback } from 'react';
 import { Breadcrumb } from 'react-breadcrumbs';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { MainRoutes } from 'shared/constants/routes';
 import DocumentTitle from 'shared/DocumentTitle';
-import { getMapiAuthUser } from 'stores/mapiauth/selectors';
+import { getMapiAuthUser } from 'stores/main/selectors';
 import styled from 'styled-components';
 
 const MapiStatus = styled.div`
@@ -19,6 +20,7 @@ const MapiStatus = styled.div`
 interface IMapiLoginPageProps {}
 
 const MapiLoginPage: React.FC<IMapiLoginPageProps> = () => {
+  const dispatch = useDispatch();
   const user = useSelector(getMapiAuthUser);
   const isLoggedIn = user !== null;
 
@@ -31,8 +33,7 @@ const MapiLoginPage: React.FC<IMapiLoginPageProps> = () => {
         const auth = MapiAuth.getInstance();
         if (isLoggedIn) {
           await auth.logout();
-          // Force a reload, so we could re-run the batched actions.
-          window.location.href = MainRoutes.Home;
+          dispatch(push(MainRoutes.Login));
         } else {
           await auth.attemptLogin();
         }
@@ -40,7 +41,7 @@ const MapiLoginPage: React.FC<IMapiLoginPageProps> = () => {
         new FlashMessage(err.message, messageType.ERROR, messageTTL.MEDIUM);
       }
     },
-    [isLoggedIn]
+    [isLoggedIn, dispatch]
   );
 
   return (
