+++ src/stores/main/constants.ts
@@ -24,3 +24,13 @@ export const REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST =
   'REQUEST_PASSWORD_RECOVERY_TOKEN_REQUEST';
 export const VERIFY_PASSWORD_RECOVERY_TOKEN = 'VERIFY_PASSWORD_RECOVERY_TOKEN';
 export const SET_NEW_PASSWORD = 'SET_NEW_PASSWORD';
+
+export const MAPI_AUTH_USER_LOAD = 'MAPI_AUTH_USER_LOAD';
+export const MAPI_AUTH_USER_LOAD_REQUEST = 'MAPI_AUTH_USER_LOAD_REQUEST';
+export const MAPI_AUTH_USER_LOAD_ERROR = 'MAPI_AUTH_USER_LOAD_ERROR';
+export const MAPI_AUTH_USER_LOAD_SUCCESS = 'MAPI_AUTH_USER_LOAD_SUCCESS';
+export const MAPI_AUTH_USER_EXPIRED = 'MAPI_AUTH_USER_EXPIRED';
+export const MAPI_AUTH_USER_EXPIRING = 'MAPI_AUTH_USER_EXPIRING';
+export const MAPI_AUTH_USER_SIGNED_OUT = 'MAPI_AUTH_USER_SIGNED_OUT';
+export const MAPI_AUTH_USER_SESSION_TERMINATED =
+  'MAPI_AUTH_USER_SESSION_TERMINATED';
