+++ src/stores/main/reducer.ts
@@ -1,6 +1,12 @@
 import produce from 'immer';
 import { Providers } from 'shared/constants';
-import { CLUSTER_SELECT } from 'stores/main/constants';
+import {
+  CLUSTER_SELECT,
+  MAPI_AUTH_USER_EXPIRED,
+  MAPI_AUTH_USER_LOAD_SUCCESS,
+  MAPI_AUTH_USER_SESSION_TERMINATED,
+  MAPI_AUTH_USER_SIGNED_OUT,
+} from 'stores/main/constants';
 import {
   GLOBAL_LOAD_ERROR,
   GLOBAL_LOAD_SUCCESS,
@@ -25,11 +31,14 @@ import {
   setUserToStorage,
 } from 'utils/localStorageUtils';
 
+import { loadMapiUser } from './actions';
+
 const initialState = (): IMainState => ({
   selectedOrganization: fetchSelectedOrganizationFromStorage(),
   selectedClusterID: null,
   firstLoadComplete: false,
   loggedInUser: fetchUserFromStorage(),
+  mapiUser: null,
   info: {
     general: {
       installation_name: '',
@@ -105,6 +114,23 @@ const makeMainReducer = () => {
 
         case CLUSTER_SELECT:
           draft.selectedClusterID = action.clusterID;
+
+          break;
+
+        case loadMapiUser().types.success as typeof MAPI_AUTH_USER_LOAD_SUCCESS:
+          draft.mapiUser = action.response;
+
+          break;
+
+        case loadMapiUser().types.error:
+          draft.mapiUser = null;
+
+          break;
+
+        case MAPI_AUTH_USER_SESSION_TERMINATED:
+        case MAPI_AUTH_USER_SIGNED_OUT:
+        case MAPI_AUTH_USER_EXPIRED:
+          draft.mapiUser = null;
 
           break;
       }
